pyproject.toml
[project]
name = "cyanomanus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "e2b-code-interpreter>=1.2.0",
    "firecrawl-py>=1.16.0",
    "googlesearch-python>=1.3.0",
    "graphrag>=2.1.0",
    "httpx[socks]>=0.28.1",
    "openai-agents>=0.0.9",
    "pandas>=2.2.3",
    "pathlib>=1.0.1",
    "pdfdeal>=1.0.2",
    "python-dotenv>=1.1.0",
    "rich>=13.9.4",
    "tavily-python>=0.5.4",
    "wikipedia>=1.4.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "langsmith[openai-agents]>=0.3.27",
]
